separate app.py in routes, utils, conf and exstensions for better readability and managment
