+++ app.py
@@ -1,438 +1,32 @@
-from flask import (
-    Flask, render_template, request, jsonify,
-    send_from_directory, redirect, make_response, url_for, flash
-)
-from flask_sqlalchemy import SQLAlchemy
-from flask_bcrypt import Bcrypt
-from flask_jwt_extended import (
-    JWTManager, create_access_token,
-    jwt_required, get_jwt_identity
-)
-from flask_mail import Mail, Message
-from itsdangerous import URLSafeTimedSerializer, SignatureExpired, BadSignature
-from dotenv import load_dotenv
-import datetime, os
-from sqlalchemy import desc
+# ╔═════════════════════════════════════════════╗
+# ║  I HAVE SPENT 500 HOURS ON THIS JUNGLE 🕳️   ║
+# ║  When I wrote this code, only I and God     ║
+# ║  knew how it worked... now only God does.   ║
+# ║  Don't waste your time here—just scroll on. ║
+# ║  If you're still reading, bump the counters:║
+# ║    hoursWasted += 1;                        ║
+# ║    sanityLevel    -= 1;                     ║
+# ╚═════════════════════════════════════════════╝
 
-# Load env
-load_dotenv()
+from flask import Flask
+from config import Config
+from extensions import init_extensions
+from routes import blueprints
 
-app = Flask(__name__, static_folder='static')
+def create_app(config_class=Config):
+    app = Flask(__name__, static_folder='static')
+    app.config.from_object(config_class)
 
-# --- Configuration ---
-app.config['SECRET_KEY'] = os.getenv('SECRET_KEY')
-app.config['SQLALCHEMY_DATABASE_URI'] = os.getenv('DATABASE_URI')
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+    # Initialize extensions
+    init_extensions(app)
 
-app.config['JWT_SECRET_KEY'] = os.getenv('JWT_SECRET_KEY')
-app.config['JWT_ACCESS_TOKEN_EXPIRES'] = datetime.timedelta(
-    minutes=int(os.getenv('JWT_EXPIRES_MINUTES', 15))
-)
-app.config['JWT_TOKEN_LOCATION'] = ['cookies']
-app.config['JWT_COOKIE_CSRF_PROTECT'] = False
-# consistent cookie name
-app.config['JWT_ACCESS_COOKIE_NAME'] = 'access_token_cookie'
+    # Register blueprints
+    for blueprint in blueprints:
+        app.register_blueprint(blueprint)
 
-# Mail settings
-app.config['MAIL_SERVER'] = os.getenv('MAIL_SERVER')
-app.config['MAIL_PORT'] = int(os.getenv('MAIL_PORT', 587))
-app.config['MAIL_USE_TLS'] = os.getenv('MAIL_USE_TLS') == 'True'
-app.config['MAIL_USERNAME'] = os.getenv('MAIL_USERNAME')
-app.config['MAIL_PASSWORD'] = os.getenv('MAIL_PASSWORD')
-app.config['MAIL_DEFAULT_SENDER'] = (
-    os.getenv('MAIL_DEFAULT_NAME'), os.getenv('MAIL_DEFAULT_EMAIL')
-)
-app.config['MAIL_DEBUG'] = True
+    return app
 
-# Extensions
-db    = SQLAlchemy(app)
-bcrypt= Bcrypt(app)
-jwt   = JWTManager(app)
-mail  = Mail(app)
-ts    = URLSafeTimedSerializer(app.config['SECRET_KEY'])
+app = create_app()
 
-# Ensure JWT sub is string
-@jwt.user_identity_loader
-def user_identity_lookup(identity):
-    return str(identity)
-
-# Models
-class User(db.Model):
-    id                 = db.Column(db.Integer, primary_key=True)
-    username           = db.Column(db.String(80), unique=True, nullable=False)
-    email              = db.Column(db.String(120), unique=True, nullable=False)
-    password_hash      = db.Column(db.String(128), nullable=False)
-    is_active          = db.Column(db.Boolean, default=False)
-    exercise_points    = db.Column(db.Integer, default=0)
-    join_date          = db.Column(db.DateTime, default=datetime.datetime.utcnow)
-    achievements_unlocked = db.Column(db.Integer, default=0)
-    exercises          = db.relationship('Exercise', backref='user', lazy=True)
-
-    def set_password(self, pw):
-        self.password_hash = bcrypt.generate_password_hash(pw).decode()
-
-    def check_password(self, pw):
-        return bcrypt.check_password_hash(self.password_hash, pw)
-
-    def get_rank(self):
-        pts = self.exercise_points
-        if pts >= 1000: return 'Master'
-        if pts >= 700:  return 'Ruby'
-        if pts >= 400:  return 'Diamond'
-        if pts >= 200:  return 'Silver'
-        return 'Bronze'
-
-    def calculate_achievements(self):
-        thresholds = [10,50,100,200,300,400,500,700,1000]
-        self.achievements_unlocked = sum(self.exercise_points>=t for t in thresholds)
-        db.session.commit()
-
-class Exercise(db.Model):
-    id            = db.Column(db.Integer, primary_key=True)
-    user_id       = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
-    exercise_type = db.Column(db.String(50), nullable=False)
-    count         = db.Column(db.Integer, nullable=False)
-    intensity     = db.Column(db.Float, nullable=False, default=1.0)
-    points        = db.Column(db.Integer, nullable=False)
-    date_added    = db.Column(db.DateTime, default=datetime.datetime.utcnow)
-
-# Helpers
-def get_current_user():
-    try:
-        uid = get_jwt_identity()
-        return User.query.get(uid) if uid else None
-    except:
-        return None
-
-# Dashboard-specific rank thresholds
-# at the top, where EXERCISE_RANKS lives:
-EXERCISE_RANKS = {
-    'pushup': [(2000,'Mythic'),(1000,'Master'),(700,'Ruby'),(400,'Diamond'),(200,'Silver'),(0,'Bronze')],
-    'situp':  [(3000,'Mythic'),(1500,'Master'),(1000,'Ruby'),(500,'Diamond'), (200,'Silver'),(0,'Bronze')],
-    'squat':  [(2500,'Mythic'),(1200,'Master'),(800,'Ruby'), (500,'Diamond'), (200,'Silver'),(0,'Bronze')],
-    'burpee': [(1000,'Mythic'),(700,'Master'), (400,'Ruby'),  (200,'Diamond'), (100,'Silver'),(0,'Bronze')],
-    'run':    [(5000,'Mythic'),(2000,'Master'),(1000,'Ruby'), (500,'Diamond'), (200,'Silver'),(0,'Bronze')],
-    'pullup': [(1500,'Mythic'),(800,'Master'), (500,'Ruby'),  (300,'Diamond'), (100,'Silver'), (0,'Bronze')],
-    'plank':  [(10000,'Mythic'),(5000,'Master'),(2000,'Ruby'),(1000,'Diamond'), (500,'Silver'), (0,'Bronze')],
-}
-
-
-# Routes
-@app.route('/')
-def home():
-    return render_template('home.html', user=get_current_user())
-
-@app.route('/auth')
-def auth_page():
-    confirmed = request.args.get('confirmed')
-    return render_template('auth.html', confirmed=bool(confirmed))
-
-@app.route('/confirm/<token>')
-def confirm_email(token):
-    try:
-        email = ts.loads(token, salt='email-confirm', max_age=int(os.getenv('CONFIRM_EXPIRATION',3600)))
-    except SignatureExpired:
-        return "Link expired",400
-    except BadSignature:
-        return "Invalid token",400
-    user = User.query.filter_by(email=email).first()
-    if user:
-        user.is_active=True; db.session.commit()
-        return redirect('/auth?confirmed=1')
-    return "User not found",404
-
-@app.route('/api/register',methods=['POST'])
-def api_register():
-    d=request.get_json();u,e,p=d.get('username'),d.get('email'),d.get('password')
-    if not all([u,e,p]): return jsonify(success=False,message='All fields required'),400
-    if User.query.filter((User.username==u)|(User.email==e)).first():
-        return jsonify(success=False,message='Username or email taken'),409
-    user=User(username=u,email=e);user.set_password(p)
-    db.session.add(user);db.session.commit()
-    token=ts.dumps(e,salt='email-confirm')
-    link=url_for('confirm_email',token=token,_external=True)
-    msg=Message("Confirm your GameFit account",recipients=[e])
-    msg.body=f"Hi {u}, confirm here:\n\n{link}"
-    try: mail.send(msg)
-    except Exception as ex: app.logger.error(f"Mail failed: {ex}")
-    return jsonify(success=True,message='Registered! Check your email.'),200
-
-@app.route('/api/login', methods=['POST'])
-def api_login():
-    data = request.get_json()
-    username = data.get('username')
-    password = data.get('password')
-    user = User.query.filter_by(username=username).first()
-
-    if not user or not user.check_password(password):
-        return jsonify(success=False, message='Invalid credentials'), 401
-    if not user.is_active:
-        return jsonify(success=False, message='Please confirm email'), 403
-
-    token = create_access_token(identity=user.id)
-    response = jsonify(success=True, message='Login successful')
-    response.set_cookie(
-        app.config['JWT_ACCESS_COOKIE_NAME'],
-        token,
-        httponly=True,
-        secure=True,
-        samesite='Lax',
-        max_age=int(os.getenv('JWT_EXPIRES_MINUTES', 15)) * 60
-    )
-    return response
-
-@app.route('/leaderboard')
-@jwt_required()
-def leaderboard():
-    user = get_current_user()
-    if not user:
-        return redirect('/auth')
-
-    top_players = User.query.order_by(desc(User.exercise_points)).limit(20).all()
-    all_users = User.query.order_by(desc(User.exercise_points)).all()
-    user_rank = next((i+1 for i, u in enumerate(all_users) if u.id == user.id), 0)
-
-    response = make_response(render_template(
-        'leaderboard.html',
-        user=user,
-        top_players=top_players,
-        user_rank=user_rank
-    ))
-    new_token = create_access_token(identity=user.id)
-    response.set_cookie(
-        app.config['JWT_ACCESS_COOKIE_NAME'],
-        new_token,
-        httponly=True,
-        secure=True,
-        samesite='Lax',
-        max_age=int(os.getenv('JWT_EXPIRES_MINUTES', 15)) * 60
-    )
-    return response
-
-
-@app.route('/achievements')
-@jwt_required()
-def achievements():
-    user = get_current_user()
-    if not user:
-        return redirect('/auth')
-
-    user.calculate_achievements()
-    response = make_response(render_template('achievements.html', user=user))
-    new_token = create_access_token(identity=user.id)
-    response.set_cookie(
-        app.config['JWT_ACCESS_COOKIE_NAME'],
-        new_token,
-        httponly=True,
-        secure=True,
-        samesite='Lax',
-        max_age=int(os.getenv('JWT_EXPIRES_MINUTES', 15)) * 60
-    )
-    return response
-
-
-@app.route('/api/log-exercise', methods=['POST'])
-@jwt_required()
-def log_exercise():
-    uid = get_jwt_identity()
-    user = User.query.get(uid)
-    if not user:
-        return jsonify(success=False, message='User not found'), 404
-
-    data = request.get_json()
-    exercise_type = data.get('exercise_type')
-    count = int(data.get('count', 0))
-    points = int(data.get('points', 0))
-    if not exercise_type or count <= 0:
-        return jsonify(success=False, message='Invalid exercise data'), 400
-
-    exercise = Exercise(
-        user_id=uid,
-        exercise_type=exercise_type,
-        count=count,
-        intensity=1.0,
-        points=points
-    )
-    user.exercise_points += points
-    db.session.add(exercise)
-    db.session.commit()
-    user.calculate_achievements()
-
-    response = make_response(jsonify(
-        success=True,
-        message=f'Exercise logged! You earned {points} points!',
-        new_total=user.exercise_points,
-        rank=user.get_rank()
-    ))
-    new_token = create_access_token(identity=uid)
-    response.set_cookie(
-        app.config['JWT_ACCESS_COOKIE_NAME'],
-        new_token,
-        httponly=True,
-        secure=True,
-        samesite='Lax',
-        max_age=int(os.getenv('JWT_EXPIRES_MINUTES', 15)) * 60
-    )
-    return response
-
-@app.route('/api/logout',methods=['POST'])
-def logout():
-    r=jsonify(success=True,message='Logged out')
-    r.delete_cookie(app.config['JWT_ACCESS_COOKIE_NAME'])
-    return r
-
-@app.route('/api/delete',methods=['DELETE'])
-@jwt_required()
-def api_delete():
-    uid=get_jwt_identity();u=User.query.get(uid)
-    if u: db.session.delete(u);db.session.commit();r=jsonify(success=True,message='Deleted');r.delete_cookie(app.config['JWT_ACCESS_COOKIE_NAME']);return r
-    return jsonify(success=False,message='Not found'),404
-
-@app.route('/profile')
-@jwt_required()
-def profile():
-    user = get_current_user()
-    if not user:
-        return redirect('/auth')
-
-    response = make_response(render_template('profile.html', user=user))
-    new_token = create_access_token(identity=user.id)
-    response.set_cookie(
-        app.config['JWT_ACCESS_COOKIE_NAME'],
-        new_token,
-        httponly=True,
-        secure=True,
-        samesite='Lax',
-        max_age=int(os.getenv('JWT_EXPIRES_MINUTES', 15)) * 60
-    )
-    return response
-
-@app.route('/profile/edit',methods=['GET','POST'])
-@jwt_required()
-def edit_profile():
-    user=get_current_user() or redirect('/auth')
-    if request.method=='POST':
-        d=request.form;nu=d.get('username').strip();ne=d.get('email').strip();npw=d.get('password').strip()
-        if not nu or not ne: return render_template('profile_edit.html',user=user,error='Username and email cannot be empty.')
-        if nu!=user.username and User.query.filter_by(username=nu).first(): return render_template('profile_edit.html',user=user,error='Username already taken')
-        if ne!=user.email and User.query.filter_by(email=ne).first(): return render_template('profile_edit.html',user=user,error='Email already in use')
-        user.username=nu;user.email=ne
-        if npw: user.set_password(npw)
-        db.session.commit()
-        return redirect('/profile')
-    return render_template('profile_edit.html',user=user)
-
-@app.route('/dashboard', methods=['GET', 'POST'])
-@jwt_required()
-def dashboard():
-    user = get_current_user() or redirect('/auth')
-
-    # Now includes icon and description
-    daily_routine = [
-        {
-            'type': 'pushup',
-            'label': 'Push-ups',
-            'icon': 'dumbbell',
-            'description': 'Targets chest, shoulders, and triceps'
-        },
-        {
-            'type': 'situp',
-            'label': 'Sit-ups',
-            'icon': 'child',
-            'description': 'Strengthens your core and abs'
-        },
-        {
-            'type': 'squat',
-            'label': 'Squats',
-            'icon': 'walking',
-            'description': 'Builds quads and glutes'
-        },
-        {
-            'type': 'pullup',
-            'label': 'Pull-ups',
-            'icon': 'person-booth',
-            'description': 'Works your back and biceps'
-        },
-        {
-            'type': 'burpee',
-            'label': 'Burpees',
-            'icon': 'running',
-            'description': 'Full-body conditioning'
-        },
-        {
-            'type': 'plank',
-            'label': 'Plank (sec)',
-            'icon': 'hourglass-start',
-            'description': 'Core stabilization and endurance'
-        },
-        {
-            'type': 'run',
-            'label': 'Running (min)',
-            'icon': 'running',
-            'description': 'Cardio, legs, and stamina'
-        },
-    ]
-
-    if request.method == 'POST':
-        logged_any = False
-        for ex in daily_routine:
-            cnt = int(request.form.get(ex['type'], 0))
-            if cnt > 0:
-                multipliers = {
-                    'pushup': 0.5,
-                    'situp': 0.3,
-                    'squat': 0.4,
-                    'pullup': 1.0,
-                    'burpee': 1.5,
-                    'plank': 0.1,
-                    'run': 2.0
-                }
-                pts = round(multipliers.get(ex['type'], 0.5) * cnt)
-                exercise = Exercise(
-                    user_id=user.id,
-                    exercise_type=ex['type'],
-                    count=cnt,
-                    intensity=1.0,
-                    points=pts
-                )
-                user.exercise_points += pts
-                db.session.add(exercise)
-                logged_any = True
-
-        if logged_any:
-            db.session.commit()
-            user.calculate_achievements()
-            flash('Exercises logged!', 'success')
-        else:
-            flash('Enter at least one exercise count.', 'warning')
-
-        return redirect('/dashboard')
-
-    # Compute per-exercise totals & ranks
-    per_ex_ranks = {}
-    for ex in daily_routine:
-        total = db.session.query(db.func.sum(Exercise.count)) \
-            .filter_by(user_id=user.id, exercise_type=ex['type']) \
-            .scalar() or 0
-
-        # determine rank based on your EXERCISE_RANKS table
-        for thresh, name in EXERCISE_RANKS[ex['type']]:
-            if total >= thresh:
-                per_ex_ranks[ex['type']] = {'total': total, 'rank': name}
-                break
-
-    return render_template(
-        'dashboard.html',
-        user=user,
-        daily_routine=daily_routine,
-        per_ex_ranks=per_ex_ranks
-    )
-
-
-@app.route('/static/<path:p>')
-def static_serve(p):
-    return send_from_directory('static',p)
-
-if __name__=='__main__':
-    with app.app_context(): db.create_all()
+if __name__ == '__main__':
     app.run(debug=True)
