+++ models/__init__.py
@@ -1,0 +1,5 @@
+from .user import User
+from .exercise import Exercise
+from .constants import EXERCISE_RANKS
+
+__all__ = ['User', 'Exercise', 'EXERCISE_RANKS']
