+++ models/exercise.py
@@ -1,0 +1,11 @@
+from datetime import datetime
+from extensions import db
+
+class Exercise(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
+    exercise_type = db.Column(db.String(50), nullable=False)
+    count = db.Column(db.Integer, nullable=False)
+    intensity = db.Column(db.Float, nullable=False, default=1.0)
+    points = db.Column(db.Integer, nullable=False)
+    date_added = db.Column(db.DateTime, default=datetime.utcnow)
